# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AddProductRequest:
    properties:
      cart_id:
        type: string
      product_id:
        type: string
      quantity:
        type: integer
    required:
    - cart_id
    - product_id
    - quantity
    type: object
  dto.Cart:
    properties:
      id:
        type: string
      lines:
        items:
          $ref: '#/definitions/dto.CartLine'
        type: array
      user:
        $ref: '#/definitions/internals_cart_controller_dto.User'
    type: object
  dto.CartLine:
    properties:
      price:
        type: number
      product:
        $ref: '#/definitions/internals_cart_controller_dto.Product'
      quantity:
        type: integer
    type: object
...
  dto.RemoveProductRequest:
    properties:
      cart_id:
        type: string
      product_id:
        type: string
    required:
    - cart_id
    - product_id
    type: object
...
  dto.UpdateCartLineRequest:
    properties:
      cart_id:
        type: string
      id:
        type: string
      product_id:
        type: string
      quantity:
        type: integer
    required:
    - cart_id
    - id
    - product_id
    - quantity
    type: object
  internals_cart_controller_dto.Product:
    properties:
      code:
        type: string
      description:
        type: string
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
  internals_cart_controller_dto.User:
    properties:
      email:
        type: string
      id:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
    type: object
host: localhost:8080
info:
  contact:
    email: anhquoc18092003@gmail.com
    name: Tran Phuoc Anh Quoc
  description: Swagger API for Go Clean Architecture.
  license:
    name: MIT
    url: https://github.com/MartinHeinz/go-project-blueprint/blob/master/LICENSE
  termsOfService: http://swagger.io/terms/
  title: Ecommerce Clean Architecture Swagger API
  version: "1.0"
paths:
...
  /carts/{userID}:
    delete:
      consumes:
      - application/json
      description: Removes a specified product from the authenticated user's shopping
        cart.
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: Product details to remove from cart
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RemoveProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Remove product from cart successfully
          schema:
            type: string
        "400":
          description: Bad Request - Invalid request parameters
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized - User ID mismatch or authentication failed
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden - User does not have the required permissions
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error - An error occurred while processing
            the request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Remove a product from the user's cart
      tags:
      - Carts
    get:
      description: Fetches the shopping cart details of the authenticated user based
        on the provided user ID.
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved the user's cart
          schema:
            $ref: '#/definitions/dto.Cart'
        "400":
          description: Bad Request - Invalid request parameters
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized - User ID mismatch or authentication failed
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden - User does not have the required permissions
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found - Cart not found for the given user ID
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error - An error occurred while processing
            the request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Retrieve the cart of a user
      tags:
      - Carts
    post:
      consumes:
      - application/json
      description: Adds a specified product to the authenticated user's shopping cart.
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: Product details to add to cart
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AddProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Add product to cart successfully
          schema:
            type: string
        "400":
          description: Bad Request - Invalid request parameters
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized - User ID mismatch or authentication failed
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden - User does not have the required permissions
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error - An error occurred while processing
            the request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Add a product to the user's cart
      tags:
      - Carts
  /carts/cart-line/{userID}:
    put:
      consumes:
      - application/json
      description: Updates the quantity or details of a specific product in the authenticated
        user's shopping cart.
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: Updated cart line details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCartLineRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Update cart successfully
          schema:
            type: string
        "400":
          description: Bad Request - Invalid request parameters
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized - User ID mismatch or authentication failed
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden - User does not have the required permissions
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error - An error occurred while processing
            the request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Update a cart line item
      tags:
      - Carts
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"